Keep shared InvertedIntegerTypeInformationEx for a follow-up PR
